chore: Add utility to take PhantomJS screenshots.
Oftentimes, a PhantomJS test will fail while it passes on other
browsers like Chrome or Firefox. This can be difficult to track
down if it is related to layout/positioning as you cannot see
what the browser looks like.

Add a new utility to capture a screenshot during your spec so
that you can more easily debug the problem.